requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
